+++ src/fst/fst.py
@@ -1810,7 +1810,7 @@ class FST:
         <<Global ROOT 0,0..0,14>.names>
 
         >>> FST('global a, b, c')[1]
-        <<Global ROOT 0,0..0,14>.names[1]>
+        'b'
 
         @public
         """